pulled out dice extension
