+++ Willowcat.CharacterGenerator.UI/ViewModel/ChartViewModel.cs
@@ -9,6 +9,7 @@ using System.Windows.Documents;
 using System.Windows.Input;
 using Willowcat.CharacterGenerator.Core;
 using Willowcat.CharacterGenerator.Core.Models;
+using Willowcat.CharacterGenerator.Model.Extension;
 using Willowcat.CharacterGenerator.UI.Commands;
 using Willowcat.CharacterGenerator.UI.Data;
 using Willowcat.CharacterGenerator.UI.Event;
